backend/requirements.txt
fastapi==0.136.0
faiss-cpu
langchain_classic==1.0.4
langchain_community==0.4.1
langchain_core==1.3.0
langchain_openai==1.1.14
langgraph==1.1.8
pandas==3.0.2
pydantic==2.13.2
python-dotenv==1.2.2
uvicorn==0.44.0